stable/tradingview-webhooks-bot/src/fast_app/src/requirements.txt
ib_async
numpy
asyncio
fastapi
python-dotenv
pytz
pathlib
requests
datetime
werkzeug
dataclasses
setuptools
waitress
uvicorn
pydantic
jinja2
python-multipart
aiofiles
tailwindcss
httpx
apscheduler
pandas
aiohttp
loguru
pandas_ta
aiosqlite
matplotlib
backtrader
feedparser
jsonpickle
kaleido